Better Result error handling
Fixes #6099.

When the backend server (Netty, Akka HTTP) discovers an error in a
result (e.g. invalid headers) the server needs to abort processing the
result and send an error result to the user. Previously, this error
handling was done with custom error handling code for the server.

This pull request changes the server behavior to use the application's
HttpErrorHandler to generate the error result instead of generating a
custom error response. Using the existing error handler makes the
error handling more standardized. It also means that the application
will be notified when an error occurs during result generation, which
will make it easier to track errors that occur in Play applications
(see #6098, #6099).

This pull request changes the signature of HttpErrorHandler's
onServerError method to take a status code parameter. Previously the
method always assumed a status code of 500. Now, the method takes a
status code parameter so that it can support additional status codes
such as 505, which is needed by this pull request.

Play's error handling code now calls the new method with the status
code. The old method is deprecated and no longer called. User code
will need to upgrade. This is documented in the Play 2.6 migration
documentation.