WIP: Changes to HTTP error handling
This pull request changes the signature of HttpErrorHandler's
onServerError method to take a status code parameter. Previously the
method always assumed a status code of 500. Now, the method takes a
status code parameter so that it can support additional status codes
such as 505, which is needed by this pull request.

Play's error handling code now calls the new method with the status
code. The old method is deprecated and no longer called. User code
will need to upgrade. This is documented in the Play 2.6 migration
documentation.